availability/serializers.py
from rest_framework import serializers
from .models import AvailabilitySlot

class AvailabilitySlotSerializer(serializers.ModelSerializer):
    class Meta:
        model = AvailabilitySlot
        fields = ['id', 'user', 'start_time', 'end_time']
        read_only_fields = ['id', 'user']
